Use hostname instead of IP for redirection.
In the Web UI, the redirection does not work for hosts where the local
interface address is not publicly accessible.  For example, with EC2 the
redirection will not work.

This change writes the Master's hostname into ZooKeeper, thereby
allowing the redirection to work with private IP addresses.

Review: https://reviews.apache.org/r/11975/