Added StateTransition to Task message.
Task now includes repeated state transitions with timestamps.  This
allows the master to perform some basic task accounting, like displaying
the start/finish time for tasks.

To support this, a timestamp was added to the TaskStatus message which
is also passed to frameworks when there are state transitions.

Review: https://reviews.apache.org/r/14434